Remove working buffer from decoding API
This was just mostly an expectation that some types will require a
working buffer for their decoding (which was indeed true for such a
primitive type as uint is), without much thinking. It turned out that
the uint (and transitively sint and float) are the only types that need
it (so far), and they will be well-served with a stack-allocated, fixed
array.